move lockrenewer out
